# Documents/WheelOfFortune_API_Spec_v1.1_fixed.yaml
openapi: 3.0.3
info:
  title: Wheel of Fortune Platform API
  version: "1.0"
  description: |
    Complete API specification for the Wheel of Fortune Customization Platform.

    **Authentication**
    - Dashboard endpoints require a Bearer JWT in the `Authorization` header.
    - Widget/public endpoints are authenticated via `embed_token` in the request body.
    - JWT access tokens expire after **15 minutes**. Use `POST /auth/refresh` to renew.

    **Base URL**
    `https://api.spinplatform.io/v1`

    **Standard Error Envelope**
    ```json
    {
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable description",
        "details": {}
      }
    }
    ```

    **Rate Limit Headers**
    All responses include: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    **Related Documents**
    - WheelOfFortune_TRD_v1.0 — Technical Requirements Document
    - WheelOfFortune_SOP_v1.0 — Standard Operating Procedure
    - WheelOfFortune_Schema_v1.0.sql — PostgreSQL Database Schema
  contact:
    name: Engineering Team
  license:
    name: Confidential — Internal Use Only
servers:
  - url: https://api.spinplatform.io/v1
    description: Production
  - url: https://staging-api.spinplatform.io/v1
    description: Staging
  - url: http://localhost:4000/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token (HS256, 15-minute expiry). Obtain from POST /auth/login."

  # ============================================================================
  # REUSABLE SCHEMAS
  # ============================================================================
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
    Timestamp:
      type: string
      format: date-time
      example: "2026-03-15T12:00:00Z"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Human readable description of the error."
            details:
              type: object
              additionalProperties: true
              example: { "field": "email", "issue": "Invalid format" }

    # --- Enums ----------------------------------------------------------------
    ClientPlan:
      type: string
      enum: [starter, growth, pro, enterprise]
    UserRole:
      type: string
      enum: [owner, admin, editor, viewer]
...
    Client:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        slug:
          type: string
        email:
          type: string
          format: email
        plan:
          $ref: '#/components/schemas/ClientPlan'
        plan_spin_limit:
          type: integer
        spins_used_this_month:
          type: integer
        billing_cycle_day:
          type: integer
          minimum: 1
          maximum: 28
        custom_domain:
          type: string
          nullable: true
        timezone:
          type: string
          example: "Asia/Kolkata"
        is_active:
          type: boolean
        created_at:
          $ref: '#/components/schemas/Timestamp'
    User:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        client_id:
          $ref: '#/components/schemas/UUID'
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        email_verified:
          type: boolean
        last_login_at:
          $ref: '#/components/schemas/Timestamp'
        created_at:
          $ref: '#/components/schemas/Timestamp'
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: "Bearer token is missing or expired."
              details: {}
...
    ValidationError:
      description: Request body failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: "One or more fields are invalid."
              details:
                email: "Must be a valid email address."
                password: "Must contain at least one uppercase letter."
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          schema:
            type: integer
          description: "Seconds until the rate limit window resets."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: TOO_MANY_REQUESTS
              message: "Rate limit exceeded. Please retry after the window resets."
              details: {}

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new client account
      description: |
        Creates a new client account and owner user in a single operation.
        Returns a JWT access token immediately. A verification email is sent
        to the provided address — the dashboard remains accessible but certain
        features require `email_verified = true`.
      operationId: registerClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [company_name, email, password, full_name]
              properties:
                company_name:
                  type: string
                  maxLength: 255
                  example: "Acme Corp"
                email:
                  type: string
                  format: email
                  example: "founder@acme.com"
                password:
                  type: string
                  minLength: 8
                  description: "Min 8 chars, at least 1 uppercase letter and 1 number."
                  example: "SecurePass1"
                full_name:
                  type: string
                  example: "Jane Doe"
      responses:
        "201":
          description: Account created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  client:
                    $ref: '#/components/schemas/Client'
                  user:
                    $ref: '#/components/schemas/User'
                  access_token:
                    type: string
                    description: "JWT (HS256, expires 15 min)."
                  refresh_token:
                    type: string
                    description: "Opaque token (expires 30 days)."
        "409":
          description: Email already registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: EMAIL_ALREADY_EXISTS
                  message: "This email address is already associated with an account."
                  details: {}
        "422":
          $ref: '#/components/responses/ValidationError'
        "429":
          description: "Registration rate limit exceeded (3 attempts / 1 hour / IP)."
          $ref: '#/components/responses/RateLimited'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain tokens
      description: |
        Validates credentials and returns a JWT access token and opaque refresh token.

        **Rate limit:** 5 failed attempts per IP per 10 minutes triggers a 30-minute lockout.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: "JWT (HS256, expires 15 min)."
                  refresh_token:
                    type: string
                    description: "Opaque token (expires 30 days)."
                  user:
                    $ref: '#/components/schemas/User'
        "401":
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: "Email or password is incorrect."
                  details: {}
        "403":
          description: Account access denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                email_not_verified:
                  summary: Email not verified
                  value:
                    error:
                      code: EMAIL_NOT_VERIFIED
                      message: "Please verify your email address before logging in."
                      details: {}
                account_suspended:
                  summary: Account suspended
                  value:
                    error:
                      code: ACCOUNT_SUSPENDED
                      message: "This account has been suspended. Contact support."
                      details: {}
        "429":
          description: "Login rate limit — 5 failed attempts in 10 min. Lockout 30 min."
          $ref: '#/components/responses/RateLimited'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: "Exchange a valid refresh token for a new access token and rotated refresh token."
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Tokens refreshed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                    description: "Rotated. Previous refresh token is invalidated."
        "401":
          description: Refresh token invalid or expired.
          $ref: '#/components/responses/Unauthorized'
...
security:
  - BearerAuth: []